requirements.txt
numpy==2.1.1
pandas==2.2.2
psycopg2-binary==2.9.9
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
pytz==2024.2
six==1.16.0
SQLAlchemy==2.0.35
typing_extensions==4.12.2
tzdata==2024.1
